Add support for dead ends
Dead ends are now calculated and kept track of. There are many possible uses for
a dead end... maybe a message or sending a user back to the last intersection?

Anyway, I have added a bool for dead ends in the block struct and a new function
gen_dead_end that determines what blocks are dead ends. This way if dead ends
are to be of use, instead of checking each time you can just access the
isDeadEnd variable.