Add "character" logic to the maze
There is now support for a "character" or a user to traverse the maze. I added
a bool to check which spot the character is in (which I may change to checking
the cx and the cy later)

Main reflects a basic implementation of it. Note that with the current
implementation it is entirely reliant on the developer to prevent the
"character" from moving in and out of walls. Keep in mind that my main does NOT
prevent the "character" from going through walls.
I don't know if I'll change this in the future.

I also moved the string creation logic of gen_main to it's own function to
accomodate the check for the "where are the available areas to move?" question
the "character" will need to ask, while still also accounting for the generation
needs of the function. It is similar to how I handled the gen_switch_case
function.

If these changes confuse you, remember that the eventual goal for this maze is
to be used in my "albino" project.