Add completion and visual logic
Finally, make use of the isSeen value. If a block is adjacent to the player,
then we'll make it visible. Print now checks to see if the block is visible
before printing it out. Move will make all eight blocks surrounding the
"character" visible.

In the getDirections function we change the logic preventing us from going
outside the maze, originally it was meant to keep the generation algorithm from
cutting into the walls. However, it no longer needs that. Now, it is just
preventing us from moving into the exit. So it is changed to only prevent
moving outside the maze.

gen_finish is added to "finalize" the maze to look good for the user. So isSeen
and the visual logic works correctly, and setting the hasPlayer/cx,cy vars to
what they should be before the user interacts with the maze.

I also changed move's type to int. It now returns a 0 when it fails (hits a
wall), a 1 when it is successful, and a 2 when the player hits the exit. These
values can be used however the developer sees fit. In my case I have the main
program ending when it sees a 2.